datatype fix in product schema

+++ dags/abdelrahman-tamer/bq_src_to_landing.py
@@ -27,7 +27,7 @@ product_schema = [
         {"name": "brand_id", "type": "INT64", "mode": "NULLABLE"},
         {"name": "category_id", "type": "INT64", "mode": "NULLABLE"},
         {"name": "product_name", "type": "STRING", "mode": "NULLABLE"},
-        {"name": "price", "type": "INT64", "mode": "NULLABLE"},
+        {"name": "price", "type": "FLOAT64", "mode": "NULLABLE"},
         {"name": "description", "type": "STRING", "mode": "NULLABLE"},
         {"name": "brand_name", "type": "STRING", "mode": "NULLABLE"},
         {"name": "category_name", "type": "STRING", "mode": "NULLABLE"},
